Auto stash before merge of "ReduceDeps" and "origin/ReduceDeps"
