+++ src/MG.Sonarr.Next.Shell.Bases/Context/CmdletContext.cs
@@ -1,4 +1,3 @@
-using MG.Http.Urls.Queries;
 using MG.Sonarr.Next.Collections.Pools;
 using MG.Sonarr.Next.Json;
 using MG.Sonarr.Next.Metadata;
@@ -139,11 +138,20 @@ namespace MG.Sonarr.Next.Shell.Context
                                    return true;
                                });
                     });
+                    //.AddGenericObjectPool<QueryParameterCollection>(builder =>
+                    //{
+                    //    builder.SetConstructor(() => new QueryParameterCollection(10))
+                    //           .SetDeconstructor(col =>
+                    //           {
+                    //               col.Clear();
+                    //               return true;
+                    //           });
+                    //});
 
             AddPool<HashSet<Wildcard>, HashSetWildcardPool>(services);
             AddPool<SortedSet<int>, SortedIntSetPool>(services);
             AddPool<PagingParameter, PagingPool>(services);
-            AddPool<Stopwatch, StopwatchPool>(services);
+            //AddPool<Stopwatch, StopwatchPool>(services);
         }
         private static void AddPool<T, TPool>(IServiceCollection services) 
             where TPool : class, IObjectPoolReturnable, IObjectPool<T>, new()
