+++ src/MG.Sonarr.Next/MG.Sonarr.Next.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -34,11 +34,13 @@
   <ItemGroup>
     <Compile Remove="Json\Converters\Spans\UtcTimeToLocalConverter.cs" />
     <Compile Remove="Strings\WildcardOld.cs" />
+    <Compile Remove="Services\Http\Queries\PagingParameterOld.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <None Include="Json\Converters\Spans\UtcTimeToLocalConverter.cs" />
     <None Include="Strings\WildcardOld.cs" />
+    <None Include="Services\Http\Queries\PagingParameterOld.cs" />
   </ItemGroup>
 
   <ItemGroup>
